refactor : start moving commands to separate files
