+++ src/main.rs
@@ -2,13 +2,13 @@ use tracing::{info, warn};
 use tracing_subscriber::{filter::LevelFilter, fmt, prelude::*, EnvFilter};
 
 use anyhow::{bail, Context};
-use clap::{builder::ArgPredicate, ArgGroup, Parser, Subcommand};
+use clap::Parser;
 use cmd_lib::{run_cmd, run_fun};
 use serde_json::json;
 
 use time::{Duration, Instant};
 
-use std::io::{BufReader, Seek, Write};
+use std::io::BufReader;
 // use std::io::{Seek, SeekFrom};
 use serde_json::Value;
 use std::path::{Path, PathBuf};
@@ -21,365 +21,8 @@ use utils::workflow_utils::*;
 
 use crate::utils::prog_utils;
 
-#[derive(Clone, Debug)]
-pub enum ReferenceType {
-    SplicedIntronic,
-    SplicedUnspliced,
-}
-
-fn ref_type_parser(s: &str) -> Result<ReferenceType, String> {
-    match s {
-        "spliced+intronic" | "splici" => Ok(ReferenceType::SplicedIntronic),
-        "spliced+unspliced" | "spliceu" => Ok(ReferenceType::SplicedUnspliced),
-        t => Err(format!("Do not recognize reference type {}", t)),
-    }
-}
-
-#[derive(Debug, Subcommand)]
-pub enum Commands {
-    /// build the (expanded) reference index
-    #[command(arg_required_else_help = true)]
-    #[command(group(
-             ArgGroup::new("reftype")
-             .required(true)
-             .args(["fasta", "ref_seq"])
-    ))]
-    Index {
-        /// specify whether an expanded reference, spliced+intronic (or splici) or spliced+unspliced (or spliceu), should be built
-        #[arg(long, help_heading="Expanded Reference Options", display_order = 1, default_value = "spliced+intronic", value_parser = ref_type_parser)]
-        ref_type: ReferenceType,
-
-        /// reference genome to be used for the expanded reference construction
-        #[arg(short, long, help_heading="Expanded Reference Options", display_order = 2, 
-              requires_ifs([
-                (ArgPredicate::IsPresent, "gtf") 
-              ]),
-              conflicts_with = "ref_seq")]
-        fasta: Option<PathBuf>,
-
-        /// reference GTF file to be used for the expanded reference construction
-        #[arg(
-            short,
-            long,
-            help_heading = "Expanded Reference Options",
-            display_order = 3,
-            requires = "fasta",
-            conflicts_with = "ref_seq"
-        )]
-        gtf: Option<PathBuf>,
-
-        /// the target read length the splici index will be built for
-        #[arg(
-            short,
-            long,
-            help_heading = "Expanded Reference Options",
-            display_order = 4,
-            requires = "fasta",
-            conflicts_with = "ref_seq"
-        )]
-        rlen: Option<u32>,
-
-        /// deduplicate identical sequences in pyroe when building an expanded reference  reference
-        #[arg(
-            long = "dedup",
-            help_heading = "Expanded Reference Options",
-            display_order = 5,
-            requires = "fasta",
-            conflicts_with = "ref_seq"
-        )]
-        dedup: bool,
-
-        /// target sequences (provide target sequences directly; avoid expanded reference construction)
-        #[arg(long, alias = "refseq", help_heading = "Direct Reference Options", display_order = 6,
-              conflicts_with_all = ["dedup", "unspliced", "spliced", "rlen", "gtf", "fasta"])]
-        ref_seq: Option<PathBuf>,
-
-        /// path to FASTA file with extra spliced sequence to add to the index
-        #[arg(
-            long,
-            help_heading = "Expanded Reference Options",
-            display_order = 7,
-            requires = "fasta",
-            conflicts_with = "ref_seq"
-        )]
-        spliced: Option<PathBuf>,
-
-        /// path to FASTA file with extra unspliced sequence to add to the index
-        #[arg(
-            long,
-            help_heading = "Expanded Reference Options",
-            display_order = 8,
-            requires = "fasta",
-            conflicts_with = "ref_seq"
-        )]
-        unspliced: Option<PathBuf>,
-
-        /// use piscem instead of salmon for indexing and mapping
-        #[arg(long, help_heading = "Piscem Index Options", display_order = 1)]
-        use_piscem: bool,
-
-        /// the value of m to be used to construct the piscem index (must be < k)
-        #[arg(
-            short = 'm',
-            long = "minimizer-length",
-            default_value_t = 19,
-            requires = "use_piscem",
-            help_heading = "Piscem Index Options",
-            display_order = 2
-        )]
-        minimizer_length: u32,
-
-        /// path to output directory (will be created if it doesn't exist)
-        #[arg(short, long, display_order = 1)]
-        output: PathBuf,
-
-        /// overwrite existing files if the output directory is already populated
-        #[arg(long, display_order = 6)]
-        overwrite: bool,
-
-        /// number of threads to use when running
-        #[arg(short, long, default_value_t = 16, display_order = 2)]
-        threads: u32,
-
-        /// the value of k to be used to construct the index
-        #[arg(
-            short = 'k',
-            long = "kmer-length",
-            default_value_t = 31,
-            display_order = 3
-        )]
-        kmer_length: u32,
-
-        /// keep duplicated identical sequences when constructing the index
-        #[arg(long, display_order = 4)]
-        keep_duplicates: bool,
-
-        /// if this flag is passed, build the sparse rather than dense index for mapping
-        #[arg(
-            short = 'p',
-            long = "sparse",
-            conflicts_with = "use_piscem",
-            display_order = 5
-        )]
-        sparse: bool,
-    },
-    /// add a new custom chemistry to geometry mapping
-    #[command(arg_required_else_help = true)]
-    AddChemistry {
-        /// the name to give the chemistry
-        #[arg(short, long)]
-        name: String,
-        /// the geometry to which the chemistry maps
-        #[arg(short, long)]
-        geometry: String,
-    },
-    /// inspect the current configuration
-    Inspect {},
-    /// quantify a sample
-    #[command(arg_required_else_help = true)]
-    #[command(group(
-            ArgGroup::new("filter")
-            .required(true)
-            .args(["knee", "unfiltered_pl", "forced_cells", "expect_cells"])
-            ))]
-    #[command(group(
-            ArgGroup::new("input-type")
-            .required(true)
-            .args(["index", "map_dir"])
-            ))]
-    Quant {
-        /// chemistry
-        #[arg(short, long)]
-        chemistry: String,
-
-        /// output directory
-        #[arg(short, long)]
-        output: PathBuf,
-
-        /// number of threads to use when running
-        #[arg(short, long, default_value_t = 16)]
-        threads: u32,
-
-        /// path to index
-        #[arg(
-            short = 'i',
-            long = "index",
-            help_heading = "Mapping Options",
-            requires_ifs([
-                (ArgPredicate::IsPresent, "reads1"),
-                (ArgPredicate::IsPresent, "reads2")
-            ])
-        )]
-        index: Option<PathBuf>,
-
-        /// comma-separated list of paths to read 1 files
-        #[arg(
-            short = '1',
-            long = "reads1",
-            help_heading = "Mapping Options",
-            value_delimiter = ',',
-            requires = "index",
-            conflicts_with = "map_dir"
-        )]
-        reads1: Option<Vec<PathBuf>>,
-
-        /// comma-separated list of paths to read 2 files
-        #[arg(
-            short = '2',
-            long = "reads2",
-            help_heading = "Mapping Options",
-            value_delimiter = ',',
-            requires = "index",
-            conflicts_with = "map_dir"
-        )]
-        reads2: Option<Vec<PathBuf>>,
-
-        /// use selective-alignment for mapping (instead of pseudoalignment with structural
-        /// constraints).
-        #[arg(short = 's', long, help_heading = "Mapping Options")]
-        use_selective_alignment: bool,
-
-        /// use piscem for mapping (requires that index points to the piscem index)
-        #[arg(long, requires = "index", help_heading = "Mapping Options")]
-        use_piscem: bool,
-
-        /// path to a mapped output directory containing a RAD file to skip mapping
-        #[arg(long = "map-dir", conflicts_with_all = ["index", "reads1", "reads2"], help_heading = "Mapping Options")]
-        map_dir: Option<PathBuf>,
-
-        /// use knee filtering mode
-        #[arg(short, long, help_heading = "Permit List Generation Options")]
-        knee: bool,
-
-        /// use unfiltered permit list
-        #[arg(short, long, help_heading = "Permit List Generation Options")]
-        unfiltered_pl: Option<Option<PathBuf>>,
-
-        /// use forced number of cells
-        #[arg(short, long, help_heading = "Permit List Generation Options")]
-        forced_cells: Option<usize>,
-
-        /// use a filtered, explicit permit list
-        #[arg(short = 'x', long, help_heading = "Permit List Generation Options")]
-        explicit_pl: Option<PathBuf>,
-
-        /// use expected number of cells
-        #[arg(short, long, help_heading = "Permit List Generation Options")]
-        expect_cells: Option<usize>,
-
-        /// The expected direction/orientation of alignments in the chemistry being processed. If
-        /// not provided, will default to `fw` for 10xv2/10xv3, otherwise `both`.
-        #[arg(short = 'd', long, help_heading="Permit List Generation Options", value_parser = clap::builder::PossibleValuesParser::new(["fw", "rc", "both"]))]
-        expected_ori: Option<String>,
-
-        /// minimum read count threshold for a cell to be retained/processed; only used with --unfiltered-pl
-        #[arg(
-            long,
-            help_heading = "Permit List Generation Options",
-            default_value_t = 10
-        )]
-        min_reads: usize,
-
-        /// transcript to gene map
-        #[arg(short = 'm', long, help_heading = "UMI Resolution Options")]
-        t2g_map: Option<PathBuf>,
-
-        /// resolution mode
-        #[arg(short, long, help_heading = "UMI Resolution Options", value_parser = clap::builder::PossibleValuesParser::new(["cr-like", "cr-like-em", "parsimony", "parsimony-em", "parsimony-gene", "parsimony-gene-em"]))]
-        resolution: String,
-    },
-    /// set paths to the programs that simpleaf will use
-    SetPaths {
-        /// path to salmon to use
-        #[arg(short, long)]
-        salmon: Option<PathBuf>,
-        /// path to piscem to use
-        #[arg(short, long)]
-        piscem: Option<PathBuf>,
-        /// path to alein-fry to use
-        #[arg(short, long)]
-        alevin_fry: Option<PathBuf>,
-        /// path to pyroe to use
-        #[arg(short = 'r', long)]
-        pyroe: Option<PathBuf>,
-    },
-
-    GetWorkflowConfig {
-        /// path to output configuration file, the directory will be created if it doesn't exist
-        #[arg(short, long)]
-        output: PathBuf,
-
-        /// name of the queried workflow.
-        #[arg(short, long)]
-        workflow: String,
-        // only write the essential information without any instructions
-        // #[arg(short, long)]
-        // essential_only: bool,
-    },
-
-    #[command(group(
-        ArgGroup::new("workflow file")
-        .required(true)
-        .args(["config_path", "workflow_path"])
-        ))]
-    /// parse the input configuration/workflow files and execute the corresponding workflow(s).
-    Workflow {
-        /// path to a simpleaf workflow configuration file.
-        #[arg(short, long, display_order = 1, help_heading = "Workflow File")]
-        config_path: Option<PathBuf>,
-
-        /// path to a simpleaf complete workflow JSON file.
-        #[arg(short, long, display_order = 2, help_heading = "Workflow File")]
-        workflow_path: Option<PathBuf>,
-
-        /// output directory for log files and the workflow outputs that have no explicit output directory.
-        #[arg(short, long, display_order = 3)]
-        output: PathBuf,
-
-        /// return after parsing the wofklow config or JSON file without executing the commands.
-        #[arg(short, long, display_order = 4, conflicts_with_all=["start_at", "resume"])]
-        no_execution: bool,
-
-        /// Start the execution from a specific step. All previous steps will be ignored.  
-        #[arg(
-            short,
-            long,
-            default_value_t = 1,
-            display_order = 5,
-            help_heading = "Start Step"
-        )]
-        start_at: isize,
-        // TODO: add a --resume arg which reads the log at starts at the step that failed in the previous run
-        /// resume execution from the termination step of a previous run. To use this flag, the output directory must contains the log file from a previous run.
-        #[arg(
-            short,
-            long,
-            conflicts_with = "start_at",
-            display_order = 6,
-            help_heading = "Start Step"
-        )]
-        resume: bool,
-
-        /// comma separated library search paths when processing the (custom) workflow configuration file. (right-most wins)
-        #[arg(
-            short,
-            long,
-            conflicts_with = "workflow_path",
-            display_order = 7,
-            value_delimiter = ','
-        )]
-        lib_paths: Option<Vec<PathBuf>>,
-
-        /// comma separated integers indicating which steps (commands) will be skipped during the execution.
-        #[arg(
-            long,
-            conflicts_with = "workflow_path",
-            display_order = 7,
-            value_delimiter = ','
-        )]
-        skip_step: Option<Vec<isize>>,
-    },
-}
+mod simpleaf_commands;
+use simpleaf_commands::*;
 
 /// simplifying alevin-fry workflows
 #[derive(Debug, Parser)]
@@ -867,94 +510,7 @@ fn build_ref_and_index(af_home_path: &Path, index_args: Commands) -> anyhow::Res
     Ok(())
 }
 
-fn inspect_simpleaf(af_home_path: PathBuf) -> anyhow::Result<()> {
-    // Read the JSON contents of the file as an instance of `User`.
-    let v: Value = inspect_af_home(af_home_path.as_path())?;
-    println!(
-        "\n----- simpleaf info -----\n{}",
-        serde_json::to_string_pretty(&v).unwrap()
-    );
 
-    // do we have a custom chemistry file
-    let custom_chem_p = af_home_path.join("custom_chemistries.json");
-    if custom_chem_p.is_file() {
-        println!(
-            "\nCustom chemistries exist at path : {}\n----- custom chemistries -----\n",
-            custom_chem_p.display()
-        );
-        // parse the custom chemistry json file
-        let custom_chem_file = std::fs::File::open(&custom_chem_p).with_context({
-            || {
-                format!(
-                    "couldn't open the custom chemistry file {}",
-                    custom_chem_p.display()
-                )
-            }
-        })?;
-        let custom_chem_reader = BufReader::new(custom_chem_file);
-        let v: Value = serde_json::from_reader(custom_chem_reader)?;
-        println!("{}", serde_json::to_string_pretty(&v).unwrap());
-    }
-    Ok(())
-}
-
-fn add_chemistry(af_home_path: PathBuf, add_chem_cmd: Commands) -> anyhow::Result<()> {
-    match add_chem_cmd {
-        Commands::AddChemistry { name, geometry } => {
-            // check geometry string, if no good then
-            // propagate error.
-            let _cg = extract_geometry(&geometry)?;
-
-            // do we have a custom chemistry file
-            let custom_chem_p = af_home_path.join("custom_chemistries.json");
-
-            let mut custom_chem_file = std::fs::OpenOptions::new()
-                .read(true)
-                .write(true)
-                .create(true)
-                .open(&custom_chem_p)
-                .with_context({
-                    || {
-                        format!(
-                            "couldn't open the custom chemistry file {}",
-                            custom_chem_p.display()
-                        )
-                    }
-                })?;
-
-            let custom_chem_reader = BufReader::new(&custom_chem_file);
-            let mut v: Value = match serde_json::from_reader(custom_chem_reader) {
-                Ok(sv) => sv,
-                Err(_) => {
-                    // the file was empty so here return an empty json object
-                    json!({})
-                }
-            };
-
-            if let Some(g) = v.get_mut(&name) {
-                let gs = g.as_str().unwrap();
-                info!("chemistry {} already existed, with geometry {}; overwriting geometry specification", name, gs);
-                *g = json!(geometry);
-            } else {
-                info!("inserting chemistry {} with geometry {}", name, geometry);
-                v[name] = json!(geometry);
-            }
-
-            custom_chem_file.set_len(0)?;
-            // custom_chem_file.seek(SeekFrom::Start(0))?;
-            // suggested by cargo clippy
-            custom_chem_file.rewind()?;
-
-            custom_chem_file
-                .write_all(serde_json::to_string_pretty(&v).unwrap().as_bytes())
-                .with_context(|| format!("could not write {}", custom_chem_p.display()))?;
-        }
-        _ => {
-            bail!("unknown command");
-        }
-    }
-    Ok(())
-}
 
 fn map_and_quant(af_home_path: &Path, quant_cmd: Commands) -> anyhow::Result<()> {
     match quant_cmd {
